+++ Othello/tuple.java
@@ -2,10 +2,10 @@
 public class tuple {
 
 	Board board;
-	byte x;
-	byte y;
+	int x;
+	int y;
 
-	public tuple(Board b, byte x, byte y) {
+	public tuple(Board b, int x, int y) {
 		this.board = b;
 		this.x = x;
 		this.y = y;
